added certificates widget backend and form validation all except skills

+++ backend/controllers/projectImageDeletion.ts
@@ -3,10 +3,18 @@ import fs from 'fs';
 const removeFile = (filename: string = "") => {
     const fileName = filename;
     const directoryPath = __basedir + "/public/projectsImage/";
-    fs.unlink(directoryPath + fileName, (err) => {
-        if (err) throw err;
-        console.log("Successfully deleted!");
-    });
+    const finalpath = directoryPath + fileName
+    if (fs.existsSync(finalpath)) {
+        console.log(`The file or directory at '${finalpath}' exists.`);
+        fs.unlink(directoryPath + fileName, (err) => {
+            if (err) throw err;
+            console.log("Successfully deleted!");
+        });
+    } else {
+        console.log(`The file or directory at '${finalpath}' does not exist.`);
+        return;
+    }
+
 
 };
 export default removeFile;
